+ copy
git-svn-id: file:///home/svn/dev/moot/tags/moot-2.0.6-pre7@1874 4a515de8-beb3-11dc-84d6-2b654028afac

+++ src/libmoot/mootTokenIO.cc
@@ -480,7 +480,7 @@ void TokenWriterNative::_put_token(const mootToken &token, mootio::mostream *os)
 	    os->puts(ai->details.empty() ? ai->tag : ai->details);
 	  }
 	  if (ai->prob != 0) {
-	    os->printf(" <%g>", ai->prob);
+	    os->printf(" <%g>", static_cast<double>(ai->prob));
 	  }
 	}
     }
